frontend-internal: move OpaqueType to a new @internal/opaque package
