+++ packages/internal-opaque/src/OpaqueType.ts
